@@ -5,7 +5,7 @@ author:
 homepage:
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -24,6 +24,7 @@ dependencies:
   # Store state:
   shared_preferences: ^2.0.5
   json_annotation: ^4.0.1
+  collection: ^1.15.0-nullsafety.4
 
 dev_dependencies:
   flutter_test:
