@@ -4,16 +4,16 @@ import 'package:flutter/material.dart';
 class ColorButton extends StatelessWidget {
   /// Creates a button of color.
   const ColorButton({
-    @required this.color,
+    required this.color,
     this.onTap,
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   /// Display color value.
   final Color color;
 
   /// Tap callback.
-  final VoidCallback onTap;
+  final VoidCallback? onTap;
 
   @override
   Widget build(BuildContext context) => InkWell(
